feature: trestle merge -l
* move functions to core/utils to solve circular dependencies

* initial list of contextual_path is not mandatory

* make initial list of get_contextual_path function not mandatory
